Update dependencies in composer.json
The versions of the dependencies listed in the composer.json file were updated. This includes both essential components of the application, such as "spatie/laravel-package-tools", and development tools, like "larastan/larastan" or "pestphp/pest". The changes aim to ensure the application is utilizing the latest, most stable versions of these dependencies for optimal performance and stability.